leader评分基本完成
leader评分基本完成

+++ src/main/java/com/lg/lg/entity/LgScoredetails.java
@@ -28,7 +28,7 @@ public class LgScoredetails extends BaseEntity {
   private BigDecimal amisAmount;
   @TableField(value = "finishedAmount")
   private BigDecimal finishedAmount;
-  private long score;
+  private BigDecimal score;
   @TableField(value = "comfirmTime")
   private LocalDateTime comfirmTime;
   //权重
